Add transition animation to the navigation
Note that this required updating the Voyager dependency from rc05
to rc08, because the older RC did not support our version of
Compose Multiplatform. The newer RC deprecated coroutineScope and
replaced it with screenModelScope, so the ViewModel code had to be
updated.
This is a common pattern in Compose Multiplatform - the iOS support
is still in Alpha and a lot of the popular libraries still haven't
reached a stable 1.0 version. Using Compose Multiplatform on Android
and iOS often means having to deal with "bleeding edge" libraries.